app/requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-jsonapi==4.3.0
djangorestframework-simplejwt==5.0.0
flake8==4.0.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
kombu==5.2.2
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.2
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.5.2
python-dotenv==0.19.2
pytz==2021.3
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
